+++ src/components/ProductsOption/NoOption.tsx
@@ -28,7 +28,7 @@ export default function NoOption({
   }
 
   useEffect(() => {
-    const newCount = getCountByOptionsId(optionCounts[0].optionsId)
+    const newCount = getCountByOptionsId(optionCounts[0]?.optionsId)
     setCount(newCount)
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [optionCounts])
@@ -56,7 +56,7 @@ export default function NoOption({
             {productData!.name}
           </div>
           <div className="flex justify-between items-center mt-3">
-            <UpdateCount optionsId={optionCounts[0].optionsId} count={count} />
+            <UpdateCount optionsId={optionCounts[0]?.optionsId} count={count} />
             <div className="leading-5 text-[color:var(--m-colors-gray900)] tracking-[-0.5px] text-lg font-semibold">
               {(salePrice * count).toLocaleString('ko-KR')}원
             </div>
