+++ src/components/products/BottomBtn.tsx
@@ -5,7 +5,7 @@ import { useRecoilState, useRecoilValue, useResetRecoilState } from 'recoil'
 import { isOptionToastState, postOptionsIdCountAtom } from '@/states/optionAtom'
 import { OptionCategoryType } from '@/types/OptionType'
 import { CardDetailType, DiscountType } from '@/types/productDataType'
-import { addCart } from '@/utils/addCartApi'
+import { addCart } from '@/utils/addCartMemberApi'
 import LikeBtn from '../Buttons/LikeBtn'
 import NoOption from '../ProductsOption/NoOption'
 import OptionCheck from '../ProductsOption/OptionCheck'
